initial upload
A test file based off of the tutorial on https://randomnerdtutorials.com/esp32-esp8266-mysql-database-php/
This code is for testing the sending of data from an ESP8266 microcontroller board to a LAMP server. It has been modified to post a value toggled by the built-in "FLASH" button on the NodeMCU and a value based off of a potentiometer hooked up to the analog pin. The sketch has also been modified so that the values from the button and potentiometer can be read at a faster rate than the data will be posted to the HTTP server.